Repackage fsc
Move fsc code to its own package, with
the expectation that it will retire
from the main repository and move to
a condo in Florida.
