+++ src/compiler/scala/tools/nsc/fsc/FscSettings.scala
@@ -3,12 +3,11 @@
  * @author  Paul Phillips
  */
 
-package scala.tools
-package nsc
-package settings
+package scala.tools.nsc.fsc
 
-import util.ClassPath
-import io.{ Path, AbstractFile }
+import scala.tools.nsc.Settings
+import scala.tools.nsc.util.ClassPath
+import scala.reflect.io.{ Path, AbstractFile }
 
 class FscSettings(error: String => Unit) extends Settings(error) {
   outer =>
@@ -23,10 +22,10 @@ class FscSettings(error: String => Unit) extends Settings(error) {
   val shutdown     = BooleanSetting("-shutdown", "Shutdown compile server")
   val server       = StringSetting ("-server",   "hostname:portnumber", "Specify compile server socket", "")
   val port         = IntSetting    ("-port",     "Search and start compile server in given port only",
-  		                                      0, Some((0, Int.MaxValue)), (_: String) => None)
+                                       0, Some((0, Int.MaxValue)), (_: String) => None)
   val preferIPv4   = BooleanSetting("-ipv4",     "Use IPv4 rather than IPv6 for the server socket")
   val idleMins     = IntSetting    ("-max-idle", "Set idle timeout in minutes for fsc (use 0 for no timeout)",
-                                              30, Some((0, Int.MaxValue)), (_: String) => None)
+                                       30, Some((0, Int.MaxValue)), (_: String) => None)
 
   // For improved help output, separating fsc options from the others.
   def fscSpecific = Set[Settings#Setting](
