+++ src/compiler/scala/tools/nsc/fsc/Socket.scala
@@ -3,13 +3,13 @@
  * @author  Paul Phillips
  */
 
-package scala.tools.nsc
-package io
+package scala.tools.nsc.fsc
 
-import java.io.{ IOException, InputStreamReader, BufferedReader, PrintWriter, Closeable }
-import java.io.{ BufferedOutputStream, BufferedReader }
-import java.net.{ InetAddress, Socket => JSocket }
+import java.io.{IOException, InputStreamReader, BufferedReader, PrintWriter, Closeable}
+import java.io.{BufferedOutputStream, BufferedReader}
+import java.net.{InetAddress, Socket => JSocket}
 import scala.io.Codec
+import scala.reflect.io.Streamable
 
 /** A skeletal only-as-much-as-I-need Socket wrapper.
  */
